+++ en-us/tutorials/Invoke-PSBasicsTutorial.ps1
@@ -11,6 +11,7 @@ if ($isWindows) {
 else {
     $root = "/"
 }
+
 $title = "$($titleStyle)Essential PowerShell Fundamentals$($reset)"
 
 #region content