+++ CHANGELOG.md
@@ -4,6 +4,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Fixed
+- Changed tutorial prompt to use Y as the continuation character for all cultures.
+- Fixed localization issues on non-Windows systems where no culture is detected. [[Issue #6](https://github.com/jdhitsolutions/PSIntro/issues/6)]
 
 ## [0.8.0] - 2025-06-16
 
@@ -15,7 +18,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - Moved tutorials to culture-specific folder and updated code to use culture-aware versions.
-- Updated `README.md`.
+- Updated `README.md`. [[Issue #5](https://github.com/jdhitsolutions/PSIntro/issues/5)]
 
 ## [0.7.0] - 2025-06-11